# src/main/java/com/ordersmicroservice/orders_microservice/dto/CountryDto.java
package com.ordersmicroservice.orders_microservice.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class CountryDto {
    private Long id;
    private String name;
    private Float tax;
    private String prefix;
    private String timeZone;
}