# src/main/java/com/ordersmicroservice/orders_microservice/dto/OrderDto.java
package com.ordersmicroservice.orders_microservice.dto;

import com.ordersmicroservice.orders_microservice.models.Address;
import jakarta.persistence.EnumType;
import jakarta.persistence.Enumerated;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class OrderDto {
    private Long id;
    private Long userId;
    private Long cartId;
    private Address fromAddress;
    private Address toAddress;
    @Enumerated(EnumType.STRING)
    private Status status;
    private String dateOrdered;
    private String dateDelivered;
}