# src/main/java/com/ordersmicroservice/orders_microservice/dto/UserDto.java
package com.ordersmicroservice.orders_microservice.dto;

import com.ordersmicroservice.orders_microservice.models.Address;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class UserDto {
    private Long id;
    private String name;
    private String lastName;
    private String email;
    private String password;
    private Integer fidelityPoints;
    private String phone;
    private Address address;
    private CountryDto country;
}